backend/requirements.txt
# Django ve temel bağımlılıklar
Django==4.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# CORS desteği
django-cors-headers==4.3.1

# Resim işleme
Pillow==10.0.1

# Production server
gunicorn==21.2.0
gevent==23.9.1

# Veritabanı (Production için)
psycopg2-binary==2.9.7

# Test ve geliştirme araçları
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0

# Kod kalitesi
black==23.9.1
isort==5.12.0
flake8==6.1.0

# Güvenlik
django-environ==0.11.2